Expand macros in version
The version for the rpm is extracted from the spec file, but it
may contain macros which are not expanded by git-build-rpm.

This commit tries to get version from the rpmspec command which is
available in newer RPM versions:

  rpmspec -q --qf '%{version}' <specfile>

It will fall back to the old behaviour in case it fails, because
rpmspec is only shipped with rpm >= 4.9.

Additional it checks for unexpanded macros in the version string and
aborts build if macros are found.